requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
joblib
requests
paramiko
boto3
Pillow
